+++ Source/Sitecore.FakeDb/Pipelines/ReleaseFakeDb/ResetSettings.cs
@@ -1,0 +1,13 @@
+﻿namespace Sitecore.FakeDb.Pipelines.ReleaseFakeDb
+{
+  using Sitecore.Configuration;
+  using Sitecore.Pipelines;
+
+  public class ResetSettings
+  {
+    public void Process(PipelineArgs args)
+    {
+      Settings.Reset();
+    }
+  }
+}